Move duplicated API assertion helpers in Account Management and Back Office to a static class in SharedKernel.Tests
