+++ application/account-management/Tests/Api/BaseApiTest.cs
@@ -1,19 +1,11 @@
-using System.Net;
 using System.Net.Http.Headers;
-using System.Text.Json;
-using FluentAssertions;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using PlatformPlatform.AccountManagement.Core.Authentication.Services;
-using PlatformPlatform.SharedKernel;
-using PlatformPlatform.SharedKernel.ApiResults;
 using PlatformPlatform.SharedKernel.SinglePageApp;
 using PlatformPlatform.SharedKernel.TelemetryEvents;
 using PlatformPlatform.SharedKernel.Tests.TelemetryEvents;
-using PlatformPlatform.SharedKernel.Validation;
 
 namespace PlatformPlatform.AccountManagement.Tests.Api;
 
@@ -56,108 +48,5 @@ public abstract class BaseApiTests<TContext> : BaseTest<TContext> where TContext
     {
         _webApplicationFactory.Dispose();
         base.Dispose(disposing);
-    }
-
-    protected static void EnsureSuccessGetRequest(HttpResponseMessage response)
-    {
-        response.EnsureSuccessStatusCode();
-        response.Content.Headers.ContentType!.MediaType.Should().Be("application/json");
-        response.Headers.Location.Should().BeNull();
-    }
-
-    protected static async Task EnsureSuccessPostRequest(
-        HttpResponseMessage response,
-        string? exact = null,
-        string? startsWith = null,
-        bool hasLocation = true
-    )
-    {
-        var responseBody = await response.Content.ReadAsStringAsync();
-        responseBody.Should().BeEmpty();
-
-        response.EnsureSuccessStatusCode();
-        response.Content.Headers.ContentType.Should().BeNull();
-
-        if (hasLocation)
-        {
-            response.Headers.Location.Should().NotBeNull();
-        }
-        else
-        {
-            response.Headers.Location.Should().BeNull();
-        }
-
-        if (exact is not null)
-        {
-            response.Headers.Location!.ToString().Should().Be(exact);
-        }
-
-        if (startsWith is not null)
-        {
-            response.Headers.Location!.ToString().StartsWith(startsWith).Should().BeTrue();
-        }
-    }
-
-    protected static void EnsureSuccessWithEmptyHeaderAndLocation(HttpResponseMessage response)
-    {
-        response.EnsureSuccessStatusCode();
-        response.Content.Headers.ContentType.Should().BeNull();
-        response.Headers.Location.Should().BeNull();
-    }
-
-    protected Task EnsureErrorStatusCode(HttpResponseMessage response, HttpStatusCode statusCode, IEnumerable<ErrorDetail> expectedErrors)
-    {
-        return EnsureErrorStatusCode(response, statusCode, null, expectedErrors);
-    }
-
-    protected async Task EnsureErrorStatusCode(
-        HttpResponseMessage response,
-        HttpStatusCode statusCode,
-        string? expectedDetail,
-        IEnumerable<ErrorDetail>? expectedErrors = null,
-        bool hasTraceId = false
-    )
-    {
-        response.StatusCode.Should().Be(statusCode);
-        response.Content.Headers.ContentType!.MediaType.Should().Be("application/problem+json");
-
-        var problemDetails = await DeserializeProblemDetails(response);
-
-        problemDetails.Should().NotBeNull();
-        problemDetails!.Status.Should().Be((int)statusCode);
-        problemDetails.Type.Should().StartWith("https://tools.ietf.org/html/rfc9110#section-15.");
-        problemDetails.Title.Should().Be(ApiResult.GetHttpStatusDisplayName(statusCode));
-
-        if (expectedDetail is not null)
-        {
-            problemDetails.Detail.Should().Be(expectedDetail);
-        }
-
-        if (expectedErrors is not null)
-        {
-            var actualErrorsJson = (JsonElement)problemDetails.Extensions["Errors"]!;
-            var actualErrors = JsonSerializer.Deserialize<ErrorDetail[]>(actualErrorsJson.GetRawText(), InfrastructureCoreConfiguration.JsonSerializerOptions);
-
-            actualErrors.Should().BeEquivalentTo(expectedErrors);
-        }
-
-        if (hasTraceId)
-        {
-            problemDetails.Extensions["traceId"]!.ToString().Should().NotBeEmpty();
-        }
-    }
-
-    protected async Task<T?> DeserializeResponse<T>(HttpResponseMessage response)
-    {
-        var responseStream = await response.Content.ReadAsStreamAsync();
-
-        return await JsonSerializer.DeserializeAsync<T>(responseStream, InfrastructureCoreConfiguration.JsonSerializerOptions);
-    }
-
-    private async Task<ProblemDetails?> DeserializeProblemDetails(HttpResponseMessage response)
-    {
-        var content = await response.Content.ReadAsStringAsync();
-
-        return JsonSerializer.Deserialize<ProblemDetails>(content, InfrastructureCoreConfiguration.JsonSerializerOptions);
     }
 }
